cloud/requirements.txt
google-cloud-pubsub
google-cloud-iot
pyowm